+++ src/ska_low_mccs_spshw/tile/tile_device.py
@@ -17,6 +17,7 @@ import os.path
 import re
 import sys
 import threading
+import warnings
 from collections import defaultdict
 from dataclasses import dataclass
 from functools import reduce, wraps
@@ -4248,6 +4249,18 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
 
         :return: PLL lock state
         """
+        # NOTE: duplication of timing_pll_lock_status. This is undesired
+        # but to assist with migration to new API,
+        # we are keeping this attribute for now.
+        # It will be removed in the future and
+        # users should migrate to using timing_pll_lock_status instead.
+        warnings.warn(
+            "MccsTile's pllLocked attribute is deprecated "
+            "and will be removed in a future release. "
+            "Please use timing_pll_lock_status instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
         return self._attribute_state["pllLocked"].read()
 
     @attribute(