+++ tests/unit/tile/test_tile_device.py
@@ -6,6 +6,7 @@
 # Distributed under the terms of the BSD 3-clause new license.
 # See LICENSE for more info.
 """This module contains the tests for MccsTile."""
+
 from __future__ import annotations
 
 import copy