+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   A side effect of this is that we now raise an exception rather than Rejecting the ``On`` command when in the on state (major change)
 * [THORN-641] Fix type being pushed. It appears on pytango 10.3.0 (and before, unsure at what version this appeared) that pushing None in a list is invalid. Instead you must push nan.
 * [THORN-641] Clean up Queued to Queued invalid transition logs.
+* [THORN-593] Report Subrack Per TPM Attributes in Tile
 
 ## 12.2.2